add_new_modalities/requirements.txt
pyorthanc
python-dotenv
pyyaml